Split Kokkos_Cuda_Parallel.hpp into three files
